refecatored appnavhost into graphs
