+++ app/src/main/java/com/swahilib/navigation/graphs/ContentNavGraph.kt
@@ -1,0 +1,129 @@
+package com.swahilib.navigation.graphs
+
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavHostController
+import androidx.navigation.NavType
+import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
+import com.swahilib.core.common.utils.Routes
+import com.swahilib.core.data.repos.PrefsRepo
+import com.swahilib.core.database.model.IdiomEntity
+import com.swahilib.core.database.model.ProverbEntity
+import com.swahilib.core.database.model.SayingEntity
+import com.swahilib.core.database.model.WordEntity
+import com.swahilib.feature.advanced_search.view.screen.AdvSearchScreen
+import com.swahilib.feature.advanced_search.viewmodel.AdvSearchViewModel
+import com.swahilib.feature.daily_content.view.DailyContentHistory
+import com.swahilib.feature.daily_content.view.DailyProverbScreen
+import com.swahilib.feature.daily_content.view.DailyWordScreen
+import com.swahilib.feature.help.view.HelpScreen
+import com.swahilib.feature.home.view.screen.HomeScreen
+import com.swahilib.feature.home.viewmodel.HomeViewModel
+import com.swahilib.feature.how_it_works.view.HowItWorksScreen
+import com.swahilib.feature.idiom.view.IdiomScreen
+import com.swahilib.feature.idiom.viewmodel.IdiomViewModel
+import com.swahilib.feature.proverb.view.screen.ProverbScreen
+import com.swahilib.feature.proverb.viewmodel.ProverbViewModel
+import com.swahilib.feature.saying.view.SayingScreen
+import com.swahilib.feature.saying.viewmodel.SayingViewModel
+import com.swahilib.feature.word.view.WordScreen
+import com.swahilib.feature.word.viewmodel.WordViewModel
+
+/** Home, dictionary entry (idiom/proverb/saying/word) detail screens, search, and daily content. */
+fun NavGraphBuilder.contentGraph(
+    navController: NavHostController,
+    prefsRepo: PrefsRepo,
+    deepLinkRoute: String?,
+    onDeepLinkConsumed: () -> Unit,
+) {
+    composable(Routes.HOME) {
+        val viewModel: HomeViewModel = hiltViewModel()
+        HomeScreen(
+            viewModel = viewModel,
+            navController = navController,
+            prefsRepo = prefsRepo,
+            deepLinkRoute = deepLinkRoute,
+            onDeepLinkConsumed = onDeepLinkConsumed
+        )
+    }
+
+    composable(Routes.IDIOM) {
+        val idiom =
+            navController.previousBackStackEntry?.savedStateHandle?.get<IdiomEntity>("idiom")
+        val viewModel: IdiomViewModel = hiltViewModel()
+        IdiomScreen(
+            navController = navController,
+            viewModel = viewModel,
+            idiom = idiom,
+            prefsRepo = prefsRepo
+        )
+    }
+
+    composable(Routes.PROVERB) {
+        val proverb =
+            navController.previousBackStackEntry?.savedStateHandle?.get<ProverbEntity>("proverb")
+        val viewModel: ProverbViewModel = hiltViewModel()
+        ProverbScreen(
+            navController = navController,
+            viewModel = viewModel,
+            proverb = proverb,
+            prefsRepo = prefsRepo
+        )
+    }
+
+    composable(Routes.SAYING) {
+        val saying =
+            navController.previousBackStackEntry?.savedStateHandle?.get<SayingEntity>("saying")
+        val viewModel: SayingViewModel = hiltViewModel()
+        SayingScreen(
+            navController = navController,
+            viewModel = viewModel,
+            saying = saying,
+            prefsRepo = prefsRepo
+        )
+    }
+
+    composable(Routes.WORD) {
+        val word =
+            navController.previousBackStackEntry?.savedStateHandle?.get<WordEntity>("word")
+        val viewModel: WordViewModel = hiltViewModel()
+        WordScreen(
+            navController = navController,
+            viewModel = viewModel,
+            word = word,
+            prefsRepo = prefsRepo
+        )
+    }
+
+    composable(Routes.ADVANCED_SEARCH) {
+        val viewModel: AdvSearchViewModel = hiltViewModel()
+        AdvSearchScreen(
+            navController = navController,
+            viewModel = viewModel,
+            prefsRepo = prefsRepo
+        )
+    }
+
+    composable(Routes.HOW_IT_WORKS) { HowItWorksScreen(navController = navController) }
+    composable(Routes.HELP) { HelpScreen(navController = navController) }
+
+    composable(Routes.DAILY_WORD) {
+        DailyWordScreen(navController = navController, prefsRepo = prefsRepo)
+    }
+
+    composable(Routes.DAILY_PROVERB) {
+        DailyProverbScreen(navController = navController, prefsRepo = prefsRepo)
+    }
+
+    composable(
+        route = Routes.DAILY_CONTENT_HISTORY,
+        arguments = listOf(
+            navArgument("type") { type = NavType.StringType }
+        ),
+    ) { backStackEntry ->
+        val type = backStackEntry.arguments?.getString("type")
+            ?: Routes.DAILY_CONTENT_TYPE_WORD
+        DailyContentHistory(navController = navController, type = type)
+    }
+}
